// src/general-blood-analysis/dto/create-analysis.dto.ts
export class CreateAnalysisDto {
  esr?: number;

  plt?: number;

  rdwSd?: number;

  pdw?: number;

  mpv?: number;

  pLcr?: number;

  pct?: number;

  lymph_rel?: number;

  lymph_abs?: number;

  mono_rel?: number;

  mono_abs?: number;

  hct?: number;

  hgb?: number;

  rbc?: number;

  mcv?: number;

  mch?: number;

  mchc?: number;

  wbc?: number;

  neut_abs?: number;

  eo_abs?: number;

  baso_abs?: number;
}